(ffmpeg 7586)
git-svn-id: https://ffdshow-tryout.svn.sourceforge.net/svnroot/ffdshow-tryout/trunk@801 3b938f2f-1a1a-0410-8054-a526ea5ff92c

+++ src/ffmpeg/libavcodec/h264.c
@@ -4469,8 +4469,8 @@ static int decode_slice_header(H264Context *h){
         return -1;
     }
 
-    if(h->dequant_coeff_pps != (int)pps_id){
-        h->dequant_coeff_pps = (int)pps_id;
+    if(h->dequant_coeff_pps != pps_id){
+        h->dequant_coeff_pps = pps_id;
         init_dequant_tables(h);
     }
